+++ train/examples/verl/configs/tool.yaml
@@ -1,0 +1,28 @@
+# Simplified MCP configuration (for user modification).
+# Only need to fill in the following fields: url, Authorization, MCP_SERVERS
+# The rest will be automatically encapsulated into the mcp_config required by the framework.
+
+# Optional: Customize service name and type (default type is streamable-http)
+server_name: aworld-mcp
+# Optional: Supported types include stdio/sse/streamable-http/api/function_tool
+# Commonly used is streamable-http, when your service follows the streamable-http protocol
+# If your service is SSE, change it to sse; if it is locally executable stdio, change it to stdio
+# If it is a built-in function tool registration, use function_tool
+# If it is a simple HTTP API list/list_tools, use api
+#
+# type: streamable-http
+
+# Required: MCP service address
+url: ""
+
+# Required: Authorization, usually Bearer XXX
+Authorization: ""
+
+# Required: List of service names to be exposed for agent use (can be a string or array)
+# e.g., "browser,filesystem,terminal" or ["browser","filesystem","terminal"]
+MCP_SERVERS: ""
+
+# Optional: Timeout configuration (seconds)
+# timeout: 600
+# sse_read_timeout: 600
+# client_session_timeout_seconds: 600
