--- train/verl/__init__.py
@@ -1,2 +1,0 @@
-# coding: utf-8
-# Copyright (c) 2025 inclusionAI.
